added controller and removed endpoints from landlordcontroller
