+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/constants/UrlSegmentConstants.kt
@@ -35,4 +35,4 @@ const val FIRE_SAFETY_PATH_SEGMENT = "fire-safety-declaration"
 const val KEEP_PROPERTY_SAFE_PATH_SEGMENT = "keep-property-safe"
 const val RESPONSIBILITY_TO_TENANTS_PATH_SEGMENT = "responsibility-to-tenants"
 const val PASSCODE_ENTRY_PATH_SEGMENT = "passcode-entry"
-const val BETA_BANNER_FEEDBACK_PATH_SEGMENT = "beta-banner-feedback"
+const val FEEDBACK_PATH_SEGMENT = "feedback"
